Revert "Rerevert "Dr. Elephant Tez Support working patch (#313)" including attribution."
This reverts commit d5476c1e005b30383a64dc6b3dda9d7d34c94ffd.
