+++ app-conf/AggregatorConf.xml
@@ -34,10 +34,6 @@
     <classname>com.linkedin.drelephant.mapreduce.MapReduceMetricsAggregator</classname>
   </aggregator>
   <aggregator>
-    <applicationtype>tez</applicationtype>
-    <classname>com.linkedin.drelephant.tez.TezMetricsAggregator</classname>
-  </aggregator>
-  <aggregator>
     <applicationtype>spark</applicationtype>
     <classname>com.linkedin.drelephant.spark.SparkMetricsAggregator</classname>
     <params>
