+++ app-conf/HeuristicConf.xml
@@ -17,123 +17,6 @@
 <!-- Heuristics configurations, each heuristic will be loaded by a particular analyser -->
 <heuristics>
 
-  <!-- TEZ HEURISTICS -->
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Mapper Data Skew</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.MapperDataSkewHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <num_tasks_severity>10, 50, 100, 200</num_tasks_severity>
-      <deviation_severity>2, 4, 8, 16</deviation_severity>
-      <files_severity>1/8, 1/4, 1/2, 1</files_severity>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Mapper GC</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.MapperGCHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <exclude_jobtypes_filter>OozieLauncher</exclude_jobtypes_filter>
-      <gc_ratio_severity>0.01, 0.02, 0.03, 0.04</gc_ratio_severity>
-      <runtime_severity_in_min>5, 10, 12, 15</runtime_severity_in_min>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Mapper Time</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.MapperTimeHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <exclude_jobtypes_filter>OozieLauncher</exclude_jobtypes_filter>
-      <short_runtime_severity_in_min>10, 4, 2, 1</short_runtime_severity_in_min>
-      <long_runtime_severity_in_min>15, 30, 60, 120</long_runtime_severity_in_min>
-      <num_tasks_severity>50, 101, 500, 1000</num_tasks_severity>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Mapper Speed</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.MapperSpeedHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <exclude_jobtypes_filter>OozieLauncher</exclude_jobtypes_filter>
-      <disk_speed_severity>1/2, 1/4, 1/8, 1/32</disk_speed_severity>
-      <runtime_severity_in_min>5, 10, 15, 30</runtime_severity_in_min>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Mapper Memory</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.MapperMemoryHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <memory_ratio_severity>0.6, 0.5, 0.4, 0.3</memory_ratio_severity>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Mapper Spill</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.MapperSpillHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <spill_severity>2.01, 2.2, 2.5, 3</spill_severity>
-      <num_tasks_severity>50, 100, 500, 1000</num_tasks_severity>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Reducer Data Skew</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.ReducerDataSkewHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <num_tasks_severity>10, 50, 100, 200</num_tasks_severity>
-      <deviation_severity>2, 4, 8, 16</deviation_severity>
-      <files_severity>1/8, 1/4, 1/2, 1</files_severity>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Reducer GC</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.ReducerGCHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <gc_ratio_severity>0.01, 0.02, 0.03, 0.04</gc_ratio_severity>
-      <runtime_severity_in_min>5, 10, 12, 15</runtime_severity_in_min>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Reducer Time</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.ReducerTimeHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <short_runtime_severity_in_min>10, 4, 2, 1</short_runtime_severity_in_min>
-      <long_runtime_severity_in_min>15, 30, 60, 120</long_runtime_severity_in_min>
-      <num_tasks_severity>50, 101, 500, 1000</num_tasks_severity>
-    </params>-->
-  </heuristic>
-
-  <heuristic>
-    <applicationtype>tez</applicationtype>
-    <heuristicname>Reducer Memory</heuristicname>
-    <classname>com.linkedin.drelephant.tez.heuristics.ReducerMemoryHeuristic</classname>
-    <viewname>views.html.help.mapreduce.helpMapperSpill</viewname>
-    <!--<params>
-      <memory_ratio_severity>0.6, 0.5, 0.4, 0.3</memory_ratio_severity>
-    </params>-->
-  </heuristic>
-
   <!-- MAP-REDUCE HEURISTICS -->
 
   <heuristic>
