+++ app/com/linkedin/drelephant/util/InfoExtractor.java
@@ -20,10 +20,6 @@ import com.linkedin.drelephant.analysis.HadoopApplicationData;
 import com.linkedin.drelephant.clients.WorkflowClient;
 import com.linkedin.drelephant.configurations.scheduler.SchedulerConfiguration;
 import com.linkedin.drelephant.configurations.scheduler.SchedulerConfigurationData;
-
-import com.linkedin.drelephant.tez.data.TezApplicationData;
-import com.linkedin.drelephant.clients.WorkflowClient;
-
 import com.linkedin.drelephant.mapreduce.data.MapReduceApplicationData;
 import com.linkedin.drelephant.schedulers.Scheduler;
 import com.linkedin.drelephant.spark.data.SparkApplicationData;
@@ -120,9 +116,6 @@ public class InfoExtractor {
     } else if ( data instanceof SparkApplicationData) {
       properties = retrieveSparkProperties((SparkApplicationData) data);
     }
-    else if(data instanceof TezApplicationData){
-      properties = retrieveTezProperties((TezApplicationData) data);
-    }
     Scheduler scheduler = getSchedulerInstance(data.getAppId(), properties);
 
     if (scheduler == null) {
@@ -170,10 +163,6 @@ public class InfoExtractor {
    * @return the retrieve mapreduce properties
    */
   public static Properties retrieveMapreduceProperties(MapReduceApplicationData appData) {
-    return appData.getConf();
-  }
-
-  public static Properties retrieveTezProperties(TezApplicationData appData) {
     return appData.getConf();
   }
